+++ src/api/stats/set.rs
@@ -2,7 +2,7 @@ use crate::api::lock;
 use crate::api::stats::Stats;
 use std::sync::Mutex;
 
-pub fn add_to_stats(stats_mutex: &Mutex<Stats>, insrt_o: Option<u32>, del_o: Option<u32>, prnd_o: Option<u32>) -> Result<Stats, &'static str> {
+pub fn add_to_stats(stats_mutex: &Mutex<Stats>, insrt_o: Option<u64>, del_o: Option<u64>, prnd_o: Option<u64>) -> Result<Stats, &'static str> {
     let mut stats = lock(stats_mutex)?;
     let stats_old = stats.clone();
     if let Some(inserted) = insrt_o {
@@ -17,7 +17,7 @@ pub fn add_to_stats(stats_mutex: &Mutex<Stats>, insrt_o: Option<u32>, del_o: Opt
     Ok(stats_old)
 }
 
-pub fn replace_stats(stats_mutex: &Mutex<Stats>, insrt_o: Option<u32>, del_o: Option<u32>, prnd_o: Option<u32>) -> Result<Stats, &'static str> {
+pub fn replace_stats(stats_mutex: &Mutex<Stats>, insrt_o: Option<u64>, del_o: Option<u64>, prnd_o: Option<u64>) -> Result<Stats, &'static str> {
     let mut stats = lock(stats_mutex)?;
     let stats_old = stats.clone();
     if let Some(inserted) = insrt_o {
@@ -35,9 +35,9 @@ pub fn replace_stats(stats_mutex: &Mutex<Stats>, insrt_o: Option<u32>, del_o: Op
 pub fn handle(
     stats_mutex: &Mutex<Stats>,
     add: bool,
-    inserted: Option<u32>,
-    deleted: Option<u32>,
-    pruned: Option<u32>
+    inserted: Option<u64>,
+    deleted: Option<u64>,
+    pruned: Option<u64>
 ) -> Result<Stats, &'static str> {
     if add {
         add_to_stats(stats_mutex, inserted, deleted, pruned)
